Read exposure or outcome data

Reads in exposure data. Checks and organises columns for use with MR or enrichment tests. Infers p-values when possible from beta and se. If it is the exposure then looks up SNPs in biomaRt to get basic info.

format_data(
  dat,
  type = "exposure",
  snps = NULL,
  header = TRUE,
  phenotype_col = "Phenotype",
  snp_col = "SNP",
  beta_col = "beta",
  se_col = "se",
  eaf_col = "eaf",
  effect_allele_col = "effect_allele",
  other_allele_col = "other_allele",
  pval_col = "pval",
  units_col = "units",
  ncase_col = "ncase",
  ncontrol_col = "ncontrol",
  samplesize_col = "samplesize",
  gene_col = "gene",
  id_col = "id",
  min_pval = 1e-200,
  z_col = "z",
  info_col = "info",
  chr_col = "chr",
  pos_col = "pos",
  log_pval = FALSE
)

Arguments

dat
Data frame. Must have header with at least SNP column present.

type="exposure".
Is this the exposure or the outcome data that is being read in?

snps=NULL
SNPs to extract. If NULL then doesn't extract any and keeps all.

phenotype_col="Phenotype"
Optional column name for the column with phenotype name corresponding the the SNP. If not present then will be created with the value "Outcome"

snp_col="SNP"
Required name of column with SNP rs IDs

beta_col="beta"
Required for MR. Name of column with effect sizes

se_col="se"
Required for MR. Name of column with standard errors

eaf_col="eaf"
Required for MR. Name of column with effect allele frequency

effect_allele_col="effect_allele"
Required for MR. Name of column with effect allele. Must contain only the characters "A", "C", "T" or "G"

other_allele_col="other_allele"
Required for MR. Name of column with non effect allele. Must contain only the characters "A", "C", "T" or "G"

pval_col="pval"
Required for enrichment tests. Name of column with p-value.

units_col="units"
Optional column name for units.

ncase_col="ncase"
Optional column name for number of cases.

ncontrol_col="ncontrol"
Optional column name for number of controls.

samplesize_col="samplesize"
Optional column name for sample size.

gene_col="gene"
Optional column name for gene name.

min_pval=1e-200
Minimum allowed pval

log_pval=FALSE
The pval is -log10(P)

Value

data frame