Combine all mr results

This function combines results of mr(), mr_heterogeneity(), mr_pleiotropy_test() and mr_singlesnp() into a single data frame. It also merges the results with outcome study level characteristics in available_outcomes(). If desired it also exponentiates results (e.g. if the user wants log odds ratio converted into odds ratios with 95 percent confidence intervals). The exposure and outcome columns from the output from mr() contain both the trait names and trait ids. The combine_all_mrresults() function splits these into separate columns by default.

combine_all_mrresults(
  res,
  het,
  plt,
  sin,
  ao_slc = T,
  Exp = F,
  split.exposure = F,
  split.outcome = F
)

Arguments

res
Results from mr().

het
Results from mr_heterogeneity().

plt
Results from mr_pleiotropy_test().

sin
Results from mr_singlesnp().

ao_slc
Logical; if set to TRUE then outcome study level characteristics are retrieved from available_outcomes(). Default is TRUE.

Exp
Logical; if set to TRUE results are exponentiated. Useful if user wants log odds ratios expressed as odds ratios. Default is FALSE.

split.exposure
Logical; if set to TRUE the exposure column is split into separate columns for the exposure name and exposure ID. Default is FALSE.

split.outcome
Logical; if set to TRUE the outcome column is split into separate columns for the outcome name and outcome ID. Default is FALSE.

Value

data frame